safari-600.7-branch/LayoutTests/fast/css/replaced-element-ignore-top-bottom.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div>
  <div style="top: 0; bottom: 0">
    <img style="height: 100%" src="data:image/gif;base64,R0lGODlhAQABAIAAAAUEBAAAACwAAAAAAQABAAACAkQBADs=">
  </div>
</div>
<script>
    async_test(function(t) {
         window.onload = function() {
             t.step(function () {assert_equals(document.querySelector("img").offsetHeight, 1); t.done();});
         };
    }, "Inapplicable |top| and |bottom| specified style shouldn't affect height calculation");
</script>
